requirements.txt
streamlit
openai-whisper
google-genai
Pillow
requests
fastapi
uvicorn
python-multipart
